LayoutTests/fast/dom/NodeList/nodelist-static-having-a-bad-time.html

<!DOCTYPE HTML">
<script src="../../../resources/js-test-pre.js"></script>

<body>
<span></span>
<span></span>
<span></span>
<script>
description('This tests that static NodeList is created with appropriate indexing type when global object is having a bad time.');

var setterCalls = 0;

Object.defineProperty(Object.prototype, 42, {
    set() { setterCalls++; },
});

var nodeList = document.querySelectorAll("span");
nodeList[42] = 1

shouldBe("setterCalls", "1");

</script>
<script src="../../../resources/js-test-post.js"></script>
