LayoutTests/fast/dom/NodeList/nodelist-static-having-a-bad-time-expected.txt

This tests that static NodeList is created with appropriate indexing type when global object is having a bad time.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS setterCalls is 1
PASS successfullyParsed is true

TEST COMPLETE
